+++ realm/realm-library/src/main/cpp/io_realm_SyncSession.cpp
@@ -24,6 +24,7 @@
 
 #include "util.hpp"
 #include "jni_util/java_global_ref.hpp"
+#include "jni_util/java_local_ref.hpp"
 #include "jni_util/java_method.hpp"
 #include "jni_util/java_class.hpp"
 #include "jni_util/jni_utils.hpp"
@@ -98,9 +99,10 @@ JNIEXPORT jlong JNICALL Java_io_realm_SyncSession_nativeAddProgressListener(JNIE
             uint64_t transferred, uint64_t transferrable) {
             JNIEnv* local_env = jni_util::JniUtils::get_env(true);
 
-            auto path = to_jstring(local_env, local_realm_path);
-            local_env->CallStaticVoidMethod(java_syncmanager_class, java_notify_progress_listener, path, listener_id,
-                                            static_cast<jlong>(transferred), static_cast<jlong>(transferrable));
+            JavaLocalRef<jstring> path(local_env, to_jstring(local_env, local_realm_path));
+            local_env->CallStaticVoidMethod(java_syncmanager_class, java_notify_progress_listener, path.get(),
+                                            listener_id, static_cast<jlong>(transferred),
+                                            static_cast<jlong>(transferrable));
 
             // All exceptions will be caught on the Java side of handlers, but Errors will still end
             // up here, so we need to do something sensible with them.
@@ -111,10 +113,6 @@ JNIEXPORT jlong JNICALL Java_io_realm_SyncSession_nativeAddProgressListener(JNIE
                 local_env->ExceptionDescribe();
                 throw std::runtime_error("An unexpected Error was thrown from Java. See LogCat");
             }
-
-            // Callback happens on a thread not controlled by the JVM. So manual cleanup is
-            // required.
-            local_env->DeleteLocalRef(path);
         };
         uint64_t token = session->register_progress_notifier(callback, type, to_bool(is_streaming));
         return static_cast<jlong>(token);
@@ -155,14 +153,14 @@ JNIEXPORT jboolean JNICALL Java_io_realm_SyncSession_nativeWaitForDownloadComple
             bool listener_registered =
                 session->wait_for_download_completion([java_session_object_ref](std::error_code error) {
                     JNIEnv* env = JniUtils::get_env(true);
-                    jobject java_error_code = nullptr;
-                    jstring java_error_message = nullptr;
+                    JavaLocalRef<jobject> java_error_code;
+                    JavaLocalRef<jstring> java_error_message;
                     if (error != std::error_code{}) {
-                        java_error_code = NewLong(env, error.value());
-                        java_error_message = env->NewStringUTF(error.message().c_str());
+                        java_error_code = JavaLocalRef<jobject>(env, NewLong(env, error.value()));
+                        java_error_message = JavaLocalRef<jstring>(env, env->NewStringUTF(error.message().c_str()));
                     }
-                    env->CallVoidMethod(java_session_object_ref.get(), java_notify_result_method, java_error_code,
-                                        java_error_message);
+                    env->CallVoidMethod(java_session_object_ref.get(), java_notify_result_method, java_error_code.get(),
+                                        java_error_message.get());
                 });
 
             return to_jbool(listener_registered);