+++ realm/realm-library/src/main/cpp/jni_util/java_local_ref.hpp
@@ -33,17 +33,22 @@ static constexpr NeedToCreateLocalRef need_to_create_local_ref{};
 template <typename T>
 class JavaLocalRef {
 public:
-    // need_to_create is useful when acquire a local ref from a global weak ref.
+    inline JavaLocalRef() noexcept
+        : m_jobject(nullptr)
+        , m_env(nullptr){};
     inline JavaLocalRef(JNIEnv* env, T obj) noexcept
         : m_jobject(obj)
         , m_env(env){};
+    // need_to_create is useful when acquire a local ref from a global weak ref.
     inline JavaLocalRef(JNIEnv* env, T obj, NeedToCreateLocalRef) noexcept
         : m_jobject(env->NewLocalRef(obj))
         , m_env(env){};
 
     inline ~JavaLocalRef()
     {
-        m_env->DeleteLocalRef(m_jobject);
+        if (m_jobject) {
+            m_env->DeleteLocalRef(m_jobject);
+        }
     }
 
     JavaLocalRef& operator=(JavaLocalRef&& rhs)
@@ -54,9 +59,11 @@ public:
     }
 
     inline JavaLocalRef(JavaLocalRef&& rhs)
-            : m_env(rhs.m_env), m_jobject(rhs.m_jobject)
+        : m_jobject(rhs.m_jobject)
+        , m_env(rhs.m_env)
     {
         rhs.m_jobject = nullptr;
+        rhs.m_env = nullptr;
     }
 
     inline operator bool() const noexcept