+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 
 ### Bug Fixes
 
+* [ObjectServer] Fixed a bug which may crash when the JNI local reference limitation was reached on sync client thread.
+
 ### Internal
 
 * Upgraded to Realm Sync 1.10.1